Move functions which don't need to be in common module
